tests: Added unit test for vaccination service

+++ api/src/modules/vaccination/vaccination.service.js
@@ -1,4 +1,5 @@
 const Boom = require('@hapi/boom');
+const { Op } = require('sequelize');
 const { v4: uuidv4 } = require('uuid');
 
 const vaccinationRepository = require('./vaccination.repository');
@@ -16,7 +17,23 @@ const formatVaccination = (vaccination) => {
   };
 };
 
-const getAllVaccination = async (userId, filters) => {
+const buildFilters = (query) => {
+  const { vaccine, animalId } = query;
+
+  return {
+    ...(vaccine && { vaccine: { [Op.iLike]: `%${vaccine}%` } }),
+    ...(animalId && { animalId }),
+  };
+};
+
+const getAllVaccination = async (userId, query) => {
+  if (!userId) throw Boom.badRequest('userId was not provided');
+  if (!query || typeof query !== 'object') {
+    throw Boom.badRequest('query was not provided');
+  }
+
+  const filters = buildFilters(query);
+
   const vaccinations = await vaccinationRepository.findAllVaccinations(
     userId,
     filters,
@@ -27,21 +44,34 @@ const getAllVaccination = async (userId, filters) => {
 };
 
 const getVaccination = async (userId, vaccinationId) => {
+  if (!userId) throw Boom.badRequest('userId was not provided');
+  if (!vaccinationId) throw Boom.badRequest('vaccinationId was not provided');
+
   const vaccination = await vaccinationRepository.findOne(
     userId,
     vaccinationId,
   );
   if (!vaccination?.id) throw Boom.notFound('Vaccination does not exist');
+
   return formatVaccination(vaccination);
 };
 
 const createVaccination = async (userId, vaccinationData) => {
+  if (!userId) throw Boom.badRequest('userId was not provided');
+  if (!vaccinationData?.animalId) {
+    throw Boom.badRequest('animalId was not provided');
+  }
+  if (!vaccinationData?.vaccine && !vaccinationData?.description) {
+    throw Boom.badRequest('vaccinationData was not provided');
+  }
+
   const animal = await animalRepository.findOne(
     userId,
     vaccinationData.animalId,
   );
-  if (!animal?.id)
+  if (!animal?.id) {
     throw Boom.notFound('Animal does not exist or does not belong to the user');
+  }
 
   const vaccination = {
     id: uuidv4(),
@@ -53,12 +83,20 @@ const createVaccination = async (userId, vaccinationData) => {
   };
 
   const newVaccination = await vaccinationRepository.create(vaccination);
-  if (!newVaccination?.id)
+  if (!newVaccination?.id) {
     throw Boom.badRequest('Something went wrong creating the vaccination');
+  }
+
   return newVaccination;
 };
 
 const updateVaccination = async (userId, vaccinationId, vaccinationData) => {
+  if (!userId) throw Boom.badRequest('userId was not provided');
+  if (!vaccinationId) throw Boom.badRequest('vaccinationId was not provided');
+  if (!vaccinationData?.vaccine && !vaccinationData?.description) {
+    throw Boom.badRequest('vaccinationData was not provided');
+  }
+
   const vaccination = await getVaccination(userId, vaccinationId);
   if (!vaccination?.id) throw Boom.conflict('vaccination does not exists');
 
@@ -75,14 +113,24 @@ const updateVaccination = async (userId, vaccinationId, vaccinationData) => {
   if (updatedRows === 0) {
     throw Boom.badRequest('Something went wrong creating the vaccination');
   }
+
   return updatedVaccination;
 };
 
 const deleteVaccination = async (userId, vaccinationId) => {
+  if (!userId) throw Boom.badRequest('userId was not provided');
+  if (!vaccinationId) throw Boom.badRequest('vaccinationId was not provided');
+
   const vaccination = await getVaccination(userId, vaccinationId);
   if (!vaccination?.id) throw Boom.conflict('vaccination does not exists');
 
-  return vaccinationRepository.destroy(vaccinationId);
+  const affectedRows = await vaccinationRepository.destroy(vaccinationId);
+
+  if (affectedRows === 0) {
+    throw Boom.badRequest('Something went wrong deleting the vaccination');
+  }
+
+  return affectedRows;
 };
 
 module.exports = {
